added geolocation to automatically detect the user's location and display the weather information. also rearranged the code to make more readible
