+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack",
-		"watch": "webpack --watch"
+    "watch": "webpack --watch",
+    "start": "webpack-dev-server --open",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -19,7 +20,18 @@
   },
   "homepage": "https://github.com/macbright/weather-app#readme",
   "devDependencies": {
+    "@babel/preset-env": "^7.7.4",
+    "css-loader": "^3.2.0",
+    "file-loader": "^4.2.0",
+    "html-loader": "^0.5.5",
+    "jest": "^24.9.0",
+    "style-loader": "^1.0.0",
     "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10"
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0"
+  },
+	"dependencies": {
+    "@babel/core": "^7.7.4",
+    "babel-jest": "^24.9.0"
   }
 }
