+++ src/key.js
@@ -1,0 +1,1 @@
+export const API_KEY = '19162a83776adb533575bcaa858e87a0';
