+++ webpack.config.js
@@ -1,12 +1,45 @@
-// eslint-disable-next-line no-undef
 const path = require('path');
 
-// eslint-disable-next-line no-undef
 module.exports = {
-	entry: './src/index.js',
-	output: {
-		filename: 'main.js',
-		// eslint-disable-next-line no-undef
-		path: path.resolve(__dirname, 'dist'),
-	},
+    mode: 'development',
+    entry: './src/index.js',
+    output: {
+        filename: 'main.js',
+        path: path.resolve(__dirname, 'dist'),
+    }, module: {
+        rules: [
+            {
+                test: /\.css$/,
+                use: [
+                    'style-loader',
+                    'css-loader',
+                ],
+            },
+            // image loader
+            {
+                test: /\.(png|svg|jpg|gif)$/,
+                use: [
+                    'file-loader',
+                ],
+            },
+            // fonts
+            {
+                test: /\.(woff|woff2|eot|ttf|otf)$/,
+                use: [
+                    'file-loader',
+                ],
+            },
+            { 
+                test: /\.html$/,
+                use: [
+                    "html-loader",
+                ],
+            }
+
+        ],
+    },
+    devServer: {
+        contentBase: './dist',
+    }
+
 };
